Added function to like posts
I have implemented a feature to enable liking of posts. Upon clicking the "Like" button, a PUT request is sent, and a "like" is added to the corresponding post. Additionally, clicking the "Like" button adds a specific class to the button element. Subsequently, if the button already possesses this class, the function is terminated before its execution. This mechanism prevents multiple likes on the same post (clientside).