blogpost styling
When a user clicks on a blog post, a CSS class is dynamically added to the post element. This class causes the post to increase in size, creating a zoomed-in effect. Subsequently, when the user clicks on the post once more, the added class is removed, restoring the post to its original size. This interaction allows the user to toggle between a larger and normal view of the blog post by clicking on it.