Update post function
Users can now update their posts after they've been published. By clicking the 'Edit' button, a set of new input fields is generated. Once the modifications are complete, the user can save the updated content. This triggers an API update, and the changes promptly manifest in the browser without the need for a page refresh. Notably, if no alterations are made, the existing data is transmitted as the new values, thus ensuring that the post retains its original state..